pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "tinyeval"
version = "1.0.0"
description = "Tiny Eval - A minimal harness for evaluating LLMs via OpenAI-compatible APIs"
readme = "README.md"
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.10"
dependencies = [
    # Dataset loading (HuggingFace datasets for benchmarks)
    "datasets>=2.16.0,<4.0",
    # HTTP client
    "aiohttp",
    # Image handling (for multimodal tasks)
    "pillow",
]

[project.scripts]
tinyeval = "tinyeval:main"

[project.optional-dependencies]
dev = ["pytest", "pytest-cov", "ruff", "pre-commit"]

[tool.setuptools]
py-modules = ["tinyeval", "core"]
packages = ["tasks"]

[tool.ruff.lint]
extend-select = ["E", "F", "I", "UP"]
ignore = ["E501"]

[tool.ruff.lint.isort]
combine-as-imports = true
known-first-party = ["tinyeval", "core", "tasks"]
